MAINT: Renamed and moved format_dict to a more fitting name.
