+++ Pynetbox_Data_Uploader/tests/test_csv_utils.py
@@ -1,5 +1,5 @@
 from unittest.mock import NonCallableMock, patch
-from netbox_api.format_dict import FormatDict
+from utils.csv_to_dict import FormatDict
 import pytest
